+++ scripts/test/shared.py
@@ -415,9 +415,6 @@ SPEC_TESTS_TO_SKIP = [
 
     # Requires better support for multi-threaded tests
     'threads/wait_notify.wast',
-
-    # Non-natural alignment is invalid for atomic operations
-    'threads/atomic.wast',
 ]
 SPEC_TESTSUITE_PROPOSALS_TO_SKIP = [
 ]